Blunder (please read bellow)
I just realized as I was closing my files for this project that the readme I submitted was blank, because I made the changes but did not save them. To avoid any naming collisions with current readme here, I changed the name of this one, which is the proper Readme.
I know it's unacceptable to submit stuff late like this, but this was an honest mistake. I promise that this won't happen again, and ask for a level of leniency in this situation.